[REFACT] CouponService를 User와 Product 용으로 분리
